auto-sorter/requirements.txt
plyer==2.0.0
watchdog==2.1.6
infi.systray==1.1.0
winshell==0.6
logging==0.5.1.2